+++ cx-portal/src/components/shared/templates/Subscription/index.tsx
@@ -32,6 +32,7 @@ import {
   LoadMoreButton,
 } from 'cx-portal-shared-components'
 import {
+  AppFiltersResponse,
   SubscriptionContent,
   SubscriptionRequestBody,
 } from 'features/appSubscription/appSubscriptionApiSlice'
@@ -51,6 +52,7 @@ export enum SubscriptionTypes {
 enum FilterType {
   REQUEST = 'request',
   ACTIVE = 'active',
+  SHOWALL = 'showAll',
 }
 
 enum StatusIdType {
@@ -67,6 +69,7 @@ enum SortType {
 
 enum ViewActionEnum {
   STATUSID = 'STATUSID',
+  OFFERID = 'OFFERID',
   SORTOPTION = 'SORTOPTION',
 }
 
@@ -87,6 +90,8 @@ enum ActionKind {
   SET_SUBSCRIPTION_AND_CARD_SUBSCRIPTION = 'SET_SUBSCRIPTION_AND_CARD_SUBSCRIPTION',
   SET_PAGE_SORTING_TYPE_FETCH_ARGS = 'SET_PAGE_SORTING_TYPE_FETCH_ARGS',
   SET_PAGE_STATUS_SORTING_FETCH_ARGS = 'SET_PAGE_STATUS_SORTING_FETCH_ARGS',
+  SET_APP_FILTERS = 'SET_APP_FILTERS',
+  SET_ACTIVE_APP_FILTER = 'SET_ACTIVE_APP_FILTER',
 }
 
 type State = {
@@ -101,6 +106,8 @@ type State = {
   sortingType: string
   fetchArgs: SubscriptionRequestType | SubscriptionRequestBody
   subscriptions: SubscriptionContent[]
+  appFilters: AppFiltersResponse[]
+  activeAppFilter: string
 }
 
 type Action = {
@@ -111,19 +118,21 @@ type Action = {
 const initialState: State = {
   searchExpr: '',
   showModal: false,
-  selected: FilterType.REQUEST,
+  selected: FilterType.SHOWALL,
   sortOption: SortType.CUSTOMER,
   cardSubscriptions: [],
   serviceProviderSuccess: false,
   page: 0,
-  statusId: StatusIdType.PENDING,
+  statusId: '',
   sortingType: SortType.COMPANY_NAME_DESC,
   fetchArgs: {
     page: 0,
-    statusId: StatusIdType.PENDING,
+    statusId: '',
     sortingType: SortType.COMPANY_NAME_DESC,
   },
   subscriptions: [],
+  appFilters: [],
+  activeAppFilter: '',
 }
 
 function reducer(state: State, { type, payload }: Action) {
@@ -144,6 +153,8 @@ function reducer(state: State, { type, payload }: Action) {
       return { ...state, serviceProviderSuccess: payload }
     case ActionKind.SET_PAGE:
       return { ...state, page: payload }
+    case ActionKind.SET_STATUS_ID:
+      return { ...state, statusId: payload }
     case ActionKind.SET_SORTING_TYPE:
       return { ...state, sortingType: payload }
     case ActionKind.SET_FETCH_ARGS:
@@ -153,6 +164,7 @@ function reducer(state: State, { type, payload }: Action) {
         ...state,
         page: payload.page,
         statusId: payload.statusId,
+        offerId: payload.offerId,
         fetchArgs: payload.fetchArgs,
         sortingType: payload.sortingType,
       }
@@ -169,16 +181,21 @@ function reducer(state: State, { type, payload }: Action) {
         page: payload.page,
         fetchArgs: payload.fetchArgs,
       }
-    case ActionKind.SET_STATUS_ID:
-      return {
-        ...state,
-        statusId: payload,
-      }
     case ActionKind.SET_SUBSCRIPTION_AND_CARD_SUBSCRIPTION:
       return {
         ...state,
         subscriptions: setData(state, payload),
         cardSubscriptions: setData(state, payload),
+      }
+    case ActionKind.SET_APP_FILTERS:
+      return {
+        ...state,
+        appFilters: payload,
+      }
+    case ActionKind.SET_ACTIVE_APP_FILTER:
+      return {
+        ...state,
+        activeAppFilter: payload,
       }
     default:
       return state
@@ -208,6 +225,7 @@ const setData = (
 interface SubscriptionType {
   providerSuccessMessage: string
   fetchQuery: (Obj: SubscriptionRequestType | SubscriptionRequestBody) => any
+  fetchAppFilters?: () => any
   headline: string
   subHeading: string
   description: string
@@ -221,6 +239,7 @@ interface SubscriptionType {
   tabLabels: {
     request: string
     active: string
+    showAll: string
   }
   doNotShowAutoSetup?: boolean
   currentSuccessType: (state: RootState) => any
@@ -232,6 +251,7 @@ export default function Subscription({
   providerSuccessMessage,
   headline,
   fetchQuery,
+  fetchAppFilters,
   subHeading,
   description,
   readMore,
@@ -259,9 +279,26 @@ export default function Subscription({
       sortingType,
       fetchArgs,
       subscriptions,
+      appFilters,
+      activeAppFilter,
     },
     setState,
   ] = useReducer(reducer, initialState)
+  let appFiltersData: AppFiltersResponse[] = useMemo(() => [], [])
+  if (fetchAppFilters) {
+    const { data } = fetchAppFilters()
+    appFiltersData = data
+  }
+
+  useEffect(() => {
+    if (appFiltersData?.length) {
+      setState({
+        type: ActionKind.SET_APP_FILTERS,
+        payload: appFiltersData,
+      })
+    }
+  }, [appFiltersData, type])
+
   const { data, refetch, isFetching } = fetchQuery(fetchArgs)
   const isSuccess = useSelector(currentProviderSuccessType)
   const success: boolean = useSelector(currentSuccessType)
@@ -276,20 +313,18 @@ export default function Subscription({
   }, [data])
 
   const setView = (e: React.MouseEvent<HTMLInputElement>) => {
+    let status = ''
+    if (e.currentTarget.value === FilterType.REQUEST) {
+      status = StatusIdType.PENDING
+    } else if (e.currentTarget.value === FilterType.ACTIVE) {
+      status = StatusIdType.ACTIVE
+    }
     setState({
       type: ActionKind.SET_SORTING_TYPE,
-      payload:
-        e.currentTarget.value === FilterType.REQUEST
-          ? StatusIdType.PENDING
-          : StatusIdType.ACTIVE,
+      payload: status,
     })
     setState({ type: ActionKind.SET_SELECTED, payload: e.currentTarget.value })
-    resetCardsAndSetFetchArgs(
-      e.currentTarget.value === FilterType.REQUEST
-        ? StatusIdType.PENDING
-        : StatusIdType.ACTIVE,
-      ViewActionEnum.STATUSID
-    )
+    resetCardsAndSetFetchArgs(status, ViewActionEnum.STATUSID)
   }
 
   const resetCardsAndSetFetchArgs = (value: string, type: string) => {
@@ -302,10 +337,12 @@ export default function Subscription({
       payload: {
         page: 0,
         statusId: type === ViewActionEnum.STATUSID ? value : statusId,
+        offerId: type === ViewActionEnum.OFFERID ? value : activeAppFilter,
         sortingType: type === ViewActionEnum.SORTOPTION ? value : sortingType,
         fetchArgs: {
           page: 0,
           statusId: type === ViewActionEnum.STATUSID ? value : statusId,
+          offerId: type === ViewActionEnum.OFFERID ? value : activeAppFilter,
           sortingType: type === ViewActionEnum.SORTOPTION ? value : sortingType,
         },
       },
@@ -369,6 +406,11 @@ export default function Subscription({
       buttonValue: FilterType.ACTIVE,
       onButtonClick: setView,
     },
+    {
+      buttonText: tabLabels.showAll,
+      buttonValue: FilterType.SHOWALL,
+      onButtonClick: setView,
+    },
   ]
 
   useEffect(() => {
@@ -384,7 +426,7 @@ export default function Subscription({
             payload: expr
               ? subscriptions &&
                 subscriptions.filter((card: SubscriptionContent) =>
-                  card.serviceName.toLowerCase().includes(expr.toLowerCase())
+                  card.offerName.toLowerCase().includes(expr.toLowerCase())
                 )
               : subscriptions,
           })
@@ -461,6 +503,33 @@ export default function Subscription({
                 />
               </div>
             </div>
+            {appFilters && appFilters.length > 0 && (
+              <div className="appFilterSection">
+                {appFilters.map((app: AppFiltersResponse) => {
+                  return (
+                    <Typography
+                      className={`appName ${
+                        activeAppFilter === app.id ? 'activeFilter' : ''
+                      }`}
+                      variant="body3"
+                      onClick={() => {
+                        setState({
+                          type: ActionKind.SET_ACTIVE_APP_FILTER,
+                          payload: app.id,
+                        })
+                        resetCardsAndSetFetchArgs(
+                          app.id,
+                          ViewActionEnum.OFFERID
+                        )
+                      }}
+                      key={app.id}
+                    >
+                      {app.name}
+                    </Typography>
+                  )
+                })}
+              </div>
+            )}
             {!subscriptions ? (
               <div className="loading-progress">
                 <CircularProgress
@@ -472,8 +541,9 @@ export default function Subscription({
               </div>
             ) : (
               <SubscriptionElements
-                type={type}
                 subscriptions={cardSubscriptions}
+                isAppFilters={appFilters.length > 0}
+                type={type}
                 refetch={refetch}
               />
             )}